create
Installing an IDE
Class demonstrations are taught using Visual Studio, which is installed on all the G.32 PCs. If using your own laptops in class you are welcome to use any IDE of your choice. For example if you are on Mac we’d recommend using XCode. To practice programming outside of class you should ensure you have an IDE installed on your home computer. If you are unsure of how to install an IDE contact your tutor for help.

If you are struggling to install a compiler on your own computer online ones are also available which allow you to run and save your code (so there’s no excuse for not coding outside of class).

You can find information on how to start a project in Visual Studio or Xcode on Aula / your Github repository.